backend/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Backend Fastapi full Authentication"
authors = ["HaMinhQuan <haminhquan12c7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
certifi = "^2024.7.4"
uvicorn = "^0.30.6"
alembic = "^1.13.2"
sqlmodel = "^0.0.21"
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
python-jose = "^3.3.0"
bcrypt = "3.2.0"
passlib = "^1.7.4"
arq = "^0.26.0"
aiomysql = "^0.2.0"
arg = "^0.0.1"
python-multipart = "^0.0.9"
pyyaml = "^6.0.2"
emails = "^0.6"
jinja2 = "^3.1.4"
fastapi-limiter = "^0.1.6"
authlib = "^1.3.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
celery = "^5.4.0"
fastapi-cache2 = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
